+++ src/rex.web/src/rex/web/__init__.py
@@ -15,6 +15,7 @@ from .command import Command, Parameter
 from .csrf import (trusted, retain_csrf_token, make_csrf_meta_tag,
         make_csrf_input_tag)
 from .handle import HandleLocation, HandleFile, HandleError
+from .path import PathMask, PathMap
 from .route import MountSetting, Route
 from .secret import SecretSetting
 from .template import (get_jinja, render_to_response, HandleTemplate,