+++ src/rex.web/src/rex/web/path.py
@@ -1,3 +1,6 @@
+#
+# Copyright (c) 2014, Prometheus Research, LLC
+#
 
 
 class Symbol(object):